SendEmail View
Move SendEmail View from Categories to Tenants
