+++ backend/rest/categories/urls.py
@@ -6,14 +6,12 @@ from .views import (
     CreateCategory,
     UpdateCategory,
     DeleteCategory,
-    SendEmail
 )
 
 # URLS for categories:
 urlpatterns = [
     path('', ListCategory.as_view()),
     path('create/', CreateCategory.as_view()),
-    path('sendEmail/', SendEmail.as_view()),
     path('<pk>/', DetailCategory.as_view()),
     path('update/<pk>/', UpdateCategory.as_view()),
     path('delete/<pk>/', DeleteCategory.as_view())
