+++ backend/rest/tenant/urls.py
@@ -3,6 +3,7 @@ from django.urls import path
 from .views import (
     ListTenant,
     ListDomain,
+    SendEmail,
     CreateTenant
 )
 
@@ -10,6 +11,7 @@ urlpatterns = [
 
     path('', ListTenant.as_view()),
     path('domain/', ListDomain.as_view()),
+    path('sendEmail/', SendEmail.as_view()),
     path("create/", CreateTenant.as_view())
 
 ]
